+++ variables.tf
@@ -146,7 +146,7 @@ variable "cluster_id" {
 
 variable "image_name" {
   type        = string
-  default     = "hpcc-lsf10-scale5193-rhel88-4-0"
+  default     = "hpcc-lsf10-scale5201-rhel88-5-0"
   description = "Name of the custom image that you want to use to create virtual server instances in your IBM Cloud account to deploy the IBM Spectrum LSF cluster. By default, the automation uses a base image with additional software packages documented [here](https://cloud.ibm.com/docs/ibm-spectrum-lsf). If you would like to include your application-specific binary files, follow the instructions in [Planning for custom images](https://cloud.ibm.com/docs/vpc?topic=vpc-planning-custom-images) to create your own custom image and use that to build the IBM Spectrum LSF cluster through this offering."
 }
 
@@ -311,7 +311,7 @@ variable "spectrum_scale_enabled" {
 
 variable "scale_storage_image_name" {
   type        = string
-  default     = "hpcc-scale5193-rhel88"
+  default     = "hpcc-scale5201-rhel88"
   description = "Name of the custom image that you would like to use to create virtual machines in your IBM Cloud account to deploy the Spectrum Scale storage cluster. By default, the automation uses a base image plus the Spectrum Scale software and any other software packages that it requires. If you would like, you can follow the instructions for [Planning for custom images](https://test.cloud.ibm.com/docs/vpc?topic=vpc-planning-custom-images) to create your own custom image and use that to build the Spectrum Scale storage cluster through this offering."
 }
 
@@ -398,7 +398,7 @@ variable "scale_compute_cluster_filesystem_mountpoint" {
 
 variable "TF_WAIT_DURATION" {
   type        = string
-  default     = "180s"
+  default     = "210s"
   description = "wait duration time set for the storage and worker node to complete the entire setup"
 }
 
@@ -531,15 +531,16 @@ variable "ldap_vsi_profile" {
 
 variable "ldap_vsi_osimage_name" {
   type        = string
-  default     = "ibm-ubuntu-22-04-3-minimal-amd64-1"
+  default     = "ibm-ubuntu-22-04-4-minimal-amd64-3"
   description = "Image name to be used for provisioning the LDAP instances."
 }
 
 variable "compute_image_name" {
   type        = string
-  default     = "hpcc-lsf10-rhel88-compute-v3"
+  default     = "hpc-lsf10-rhel88-worker-v1"
   description = "Name of the custom image that you want to use to create virtual server instances in your IBM Cloud account to deploy the IBM Cloud HPC cluster dynamic compute nodes. By default, the solution uses a RHEL 8-6 OS image with additional software packages mentioned [here](https://cloud.ibm.com/docs/hpc-spectrum-LSF#create-custom-image). If you would like to include your application-specific binary files, follow the instructions in [ Planning for custom images ](https://cloud.ibm.com/docs/vpc?topic=vpc-planning-custom-images) to create your own custom image and use that to build the IBM Cloud HPC cluster through this offering."
 }
+
 variable "enable_vpc_flow_logs" {
   description = "Flag to enable VPC flow logs. If true, a flow log collector will be created."
   type        = bool